Refactor for better separation of concerns and readability
