+++ src/models/product.model.ts
