--- src/shared/types/commons.ts
@@ -1,3 +1,0 @@
-export type TokenPayload = {
-  id: string;
-};
