+++ tests/integration/auth.middleware.test.ts
@@ -1,10 +1,10 @@
 import request from "supertest";
 import { server } from "../../index";
-import { UserModel } from "../../src/domain/models/user.model";
+import { UserModel } from "../../src/models/user.model";
 
 let serverInstance: any;
 
-describe("auth middleware", () => {
+describe("auth middlewares", () => {
   let token: string;
   beforeEach(async () => {
     serverInstance = server;
